// server/src/auth/types/auth.messages.ts
export enum AuthMessages {
  EMAIL_NOT_FOUND = 'Email not found',
  PASSWORD_INCORRECT = 'Password is incorrect',
  SIGN_IN_SUCCESS = 'Sign in successfully',
  SIGN_UP_SUCCESS = 'Sign up successfully',
  GET_PROFILE_SUCCESS = 'Get profile successfully',
  UPDATE_PROFILE_SUCCESS = 'Update profile successfully',
  EMAIL_EXISTS = 'Email already exists',
}